feat: ModelType DTO, Entity, ToDomain 함수 정리
